Add writing of clinical headers
